pyproject.toml
[tool.poetry]
name = "axeratos-project"
version = "0.1.0"
description = ""
authors = ["faurdent"]
readme = "README.md"
packages = [{include = "axeratos_project"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.4"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.7"
gunicorn = "^21.2.0"
drf-yasg = {extras = ["validation"], version = "^1.21.7"}


[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
flake8 = "^6.1.0"
pre-commit = "^3.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length=120
